+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 1.12.20, 12/30/2019
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+Simplified key management for Event Node
+
+---
 ## Version 1.12.19, 12/28/2019
 
 ### Added